- change BinarySearchTree to Int

+++ app/src/main/java/com/vitorsousa/masteringdatastructures/binaryTree/Node.kt
@@ -1,6 +1,6 @@
 package com.vitorsousa.masteringdatastructures.binaryTree
 
-data class Node(val value: Byte) {
+data class Node(val value: Int) {
     var left: Node? = null
     var right: Node? = null
     override fun toString(): String {
